Compatibility work for the versioned-datastore backend
When the GBIF data is ingested (see data-importer2 repo for details) we
only pull in the id and issue field from the processed occurrences.
These are attached to records with matching occurenceID values in mongo
and then indexed into elasticsearch under the fields "gbifID" and
"gbifIssue". This then means there is no gbif table of data and
therefore to look up the gbif record version of one of our Portal
records we should use the GBIF occurrence API to do it. This is what is
now implemented in this commit.